+++ src/nvzimlib/zim_note.py
@@ -5,11 +5,12 @@ For further information see https://github.com/peter88213/
 License: GNU GPLv3 (https://www.gnu.org/licenses/gpl-3.0.en.html)
 """
 from nvzimlib.nvzim_globals import StopParsing
+from nvzimlib.nvzim_globals import ZIM_NOTE_TAG
 
 
-class ZimWikiPage:
+class ZimNote:
 
-    ZIM_PAGE_HEADER = '''Content-Type: text/x-zim-wiki
+    PAGE_HEADER = '''Content-Type: text/x-zim-wiki
 Wiki-Format: zim 0.4
 '''
     REPLACEMENTS = {
@@ -22,32 +23,58 @@ Wiki-Format: zim 0.4
         self.element = element
 
     def body(self, text):
+        """Parser callback method for a note's body text line."""
         pass
 
     def create_link(self):
+        """Add or overwrite the element's Zim note link."""
         fields = self.element.fields
-        fields['wiki-page'] = self.filePath
+        fields[ZIM_NOTE_TAG] = self.filePath
         self.element.fields = fields
 
     def fill_page(self, lines):
+        """Add page content to the lines."""
+        # To be overridden by subclasses.
         pass
 
     def from_wiki(self, text):
+        """Return text with Zim-specific formatting removed."""
         for tag in self.REPLACEMENTS:
             text = text.replace(tag, self.REPLACEMENTS[tag])
         return text
 
+    def get_h1(self, text):
+        """Return text, formatted as first level heading."""
+        return f'====== {text} ======'
+
+    def get_h2(self, text):
+        """Return text, formatted as second level heading."""
+        return f'===== {text} ====='
+
+    def get_h3(self, text):
+        """Return text, formatted as third level heading."""
+        return f'==== {text} ===='
+
     def h1(self, heading):
+        """Parser callback method for a note's first level heading."""
         if self.element.title is None:
             self.element.title = heading
 
     def h2(self, heading):
+        """Parser callback method for a note's second level heading."""
         pass
 
     def h3(self, heading):
+        """Parser callback method for a note's third level heading."""
         pass
 
+    def new_title(self):
+        """Return the title for a new note."""
+        # Override this if another title is required.
+        return self.element.title
+
     def parse_line(self, line):
+        """An event-driven line parser."""
         if line.startswith('====== '):
             heading = line.strip('= ')
             self.h1(heading)
@@ -64,6 +91,7 @@ Wiki-Format: zim 0.4
             self.body(line)
 
     def read(self):
+        """Modify the element with data read from the note file."""
         with open (self.filePath, 'r', encoding='utf-8') as f:
             text = f.read()
         lines = text.split('\n')
@@ -74,9 +102,18 @@ Wiki-Format: zim 0.4
                 return
 
     def write(self):
+        """Write the note.
+        
+        Page content:
+        - The Zim note header 
+        - A first level heading with the note title as specified by the new_title() method.
+        - Note text as specified by the fill_page() method.
+        
+        Overwrite existing note.
+        """
         lines = [
-            self.ZIM_PAGE_HEADER,
-            f'====== {self.element.title} ======\n\n',
+            self.PAGE_HEADER,
+            f'{self.get_h1(self.new_title())}\n\n',
         ]
         self.fill_page(lines)
         text = '\n'.join(lines)