+++ tools/build.py
@@ -12,7 +12,7 @@ import sys
 sys.path.insert(0, f'{os.getcwd()}/../../novelibre/tools')
 from package_builder import PackageBuilder
 
-VERSION = '0.1.0'
+VERSION = '0.2.0'
 
 
 class PluginBuilder(PackageBuilder):